requirements.txt
praw==7.7.1
google-cloud-bigquery==3.22.0
functions-framework==3.0.0
flask>=1.0,<3.0